feat(arborist): refactor arborist bin to use consisten timing/logging
This attempts to make the arborist bin script behave more like the npm
cli with regards to the handing of timing and logging.

It also adds the a `logfile` argument to write logs to a file instead of
(or in addition to) stderr. This can be helpful for benchmarking
performance of loggins or terminal display.