chore: stop publish smoke from check git clean (#8058)
Im a big proponent of scripts working the same way in ci as they would
locally. and that we should be able to run anything the ci does locally.
The publish script is used to publish npm and ws packages, but also
within a series of smoke tests. Making the flags clear it's either one
of the two paths was my goal.